requirements.txt
pycryptodome==3.23.0
pandas==2.3.3
matplotlib==3.10.8
